change plots folder
