+++ regression/scripts/aerodynamics/control_surfaces_vlm.py
@@ -13,9 +13,9 @@ import sys
 import numpy as np 
 
 import SUAVE
-from SUAVE.Core                                                     import Data, Units
-from SUAVE.Methods.Aerodynamics.Common.Fidelity_Zero.Lift           import VLM as VLM
-from SUAVE.Plots.Geometry.plot_vehicle_vlm_panelization             import plot_vehicle_vlm_panelization
+from SUAVE.Core                                                                import Data, Units
+from SUAVE.Methods.Aerodynamics.Common.Fidelity_Zero.Lift                      import VLM as VLM
+from SUAVE.Plots.Geometry.Three_Dimensional.plot_3d_vehicle_vlm_panelization   import plot_3d_vehicle_vlm_panelization
 
 sys.path.append('../Vehicles')
 
@@ -46,7 +46,7 @@ def main():
         data        = VLM(conditions, settings, geometry)
         
         plot_title  = "{}, deflection = {} degrees".format(geometry.tag, round(deflection/Units.degrees))
-        plot_vehicle_vlm_panelization(geometry, plot_control_points=False, save_filename=plot_title)        
+        plot_3d_vehicle_vlm_panelization(geometry, plot_control_points=False, save_filename=plot_title)        
         
         results.CL  = np.append(results.CL , data.CL.flatten() )
         results.CDi = np.append(results.CDi, data.CDi.flatten())
