+++ regression/scripts/aerodynamics/sideslip_and_rotation_vlm.py
@@ -13,9 +13,9 @@ import sys
 import numpy as np 
 
 import SUAVE
-from SUAVE.Core                                                     import Data, Units
-from SUAVE.Methods.Aerodynamics.Common.Fidelity_Zero.Lift           import VLM as VLM
-from SUAVE.Plots.Geometry.plot_vehicle_vlm_panelization             import plot_vehicle_vlm_panelization
+from SUAVE.Core                                                                import Data, Units
+from SUAVE.Methods.Aerodynamics.Common.Fidelity_Zero.Lift                      import VLM as VLM
+from SUAVE.Plots.Geometry.Three_Dimensional.plot_3d_vehicle_vlm_panelization   import plot_3d_vehicle_vlm_panelization
 
 sys.path.append('../Vehicles')
 
@@ -35,7 +35,7 @@ def main():
     data        = VLM(conditions, settings, geometry)
     
     plot_title  = geometry.tag
-    plot_vehicle_vlm_panelization(geometry, plot_control_points=False, save_filename=plot_title)
+    plot_3d_vehicle_vlm_panelization(geometry, plot_control_points=False, save_filename=plot_title)
 
     # save/load results
     results = Data()
