+++ trunk/SUAVE/Plots/Geometry/Common/__init__.py
@@ -1,0 +1,5 @@
+## @defgroup Plots-Geometry-Common Common
+# Description
+# @ingroup Geometry
+
+from .contour_surface_slice            import contour_surface_slice
