+++ trunk/SUAVE/Plots/Geometry/Common/contour_surface_slice.py
@@ -1,0 +1,23 @@
+## @ingroup Plots-Geometry-Common
+# contour_surface_slices.py
+#
+# Created : Mar 2020, M. Clarke
+# Modified: Apr 2020, M. Clarke
+# Modified: Jul 2020, M. Clarke
+# Modified: Jul 2021, E. Botero
+# Modified: Oct 2021, M. Clarke
+# Modified: Dec 2021, M. Clarke
+# Modified: Feb 2022, R. Erhard
+# Modified: Mar 2022, R. Erhard
+# Modified: Sep 2022, M. Clarke
+# Modified: Nov 2022, M. Clarke
+
+# ----------------------------------------------------------------------
+#  Imports
+# ---------------------------------------------------------------------- 
+import numpy as np 
+import plotly.graph_objects as go   
+
+## @ingroup Plots-Geometry-Common
+def contour_surface_slice(x,y,z,values,color_scale):
+    return go.Surface(x=x,y=y,z=z,surfacecolor=values,colorscale=color_scale, showscale=False)
