+++ trunk/SUAVE/Plots/Geometry/Three_Dimensional/__init__.py
@@ -1,0 +1,13 @@
+## @defgroup Plots-Geometry-Three_Dimensional Three_Dimensional
+# Description
+# @ingroup Geometry
+
+from .plot_3d_vehicle                  import plot_3d_vehicle
+from .plot_3d_vehicle                  import plot_3d_energy_network 
+from .plot_3d_rotor                    import plot_3d_rotor
+from .plot_3d_rotor                    import get_3d_blade_coordinates
+from .plot_3d_rotor                    import plot_3d_rotor_wake
+from .plot_3d_nacelle                  import plot_3d_nacelle 
+from .plot_3d_nacelle                  import generate_3d_nacelle_points  
+from .plot_3d_wing                     import plot_3d_wing
+from .plot_3d_vehicle_vlm_panelization import plot_3d_vehicle_vlm_panelization
