+++ trunk/SUAVE/Plots/Geometry/Two_Dimensional/__init__.py
@@ -1,0 +1,6 @@
+## @defgroup Plots-Geometry-Two_Dimensional Two_Dimensional
+# Description
+# @ingroup Geometry
+
+from .plot_airfoil                 import plot_airfoil
+from .plot_rotor                   import plot_rotor
