+++ trunk/SUAVE/Plots/Geometry/__init__.py
@@ -2,7 +2,6 @@
 # Description
 # @ingroup Plots
 
-from .plot_airfoil                  import plot_airfoil
-from .plot_rotor                    import plot_rotor
-from .plot_vehicle                  import plot_vehicle
-from .plot_vehicle_vlm_panelization import plot_vehicle_vlm_panelization
+from . import Two_Dimensional 
+from . import Three_Dimensional 
+from . import Common
